One more test

+++ config_test.go
@@ -415,16 +415,16 @@ func TestBasicConfigLoadWithUndefinedFlags(t *testing.T) {
 	assert.NotNil(t, err, "The error should not be nil")
 
 	a := Require("addend.a").Int()
-	assert.Equal(t, a, 4, "addend.a should be 4")
+	assert.Equal(t, a, int64(4), "addend.a should be 4")
 
 	b := Require("addend.b").Float()
-	assert.Equal(t, b, 2, "addend.b should be 2")
+	assert.Equal(t, b, float64(2), "addend.b should be 2")
 
 	sub := Require("subtract").Bool()
 	assert.Equal(t, sub, true, "subtract should be true")
 
 	c := float64(a) - float64(b)
-	assert.Equal(t, c, 4-2, "The operation on addend.a - addend.b should be 2")
+	assert.Equal(t, c, float64(4-2), "The operation on addend.a - addend.b should be 2")
 
 	name := Require("name").String()
 	assert.Equal(t, name, "Test", "Name should be \"Test\"")
